build simplification
 target processor (i.e. jn5148 or jn5139) is not done explicitly anymore, you
won't have to edit the platforms Makefile anymore. Compilation will be done for
the JN5148 module per default. If the GCC/Jennic toolchain version 4.2.1 is
found in the path, the compilation will take place for the JN5139.

 So you chose to compile for jn5139 by putting compiler version 4.2.1 in the
path, and compiling for JN5148 with every other version (i.e. the versions of
the two toolchains for both modules is different!)